requirements.txt
opencv==3.4.1
h5py==2.8.0
image==1.5.24
imageio==2.4.1
imutils==0.4.6
Keras==2.1.3
matplotlib==2.2.2
numpy==1.14.3
pydot==1.0.29
pydot-ng==1.0.0
scikit-image==0.14.0
scikit-learn==0.19.1
scikit-video==1.1.10
scipy==1.1.0
subprocess32==3.5.1
tensorboard-logger==0.1.0
tensorflow-gpu==1.4.0
tensorflow-tensorboard==0.4.0
tensorlayer==1.10.1
testpath==0.3.1
Theano==1.0.2
torch==0.4.0
torchvision==0.2.1
tqdm==4.23.4
